--- a/edk2toolext/invocables/edk2_pr_eval.py
+++ b/edk2toolext/invocables/edk2_pr_eval.py
@@ -198,7 +198,7 @@ class Edk2PrEval(Edk2MultiPkgAwareInvocable):
             print(self.output_count_format_string.format(pkgcount=pkgcount))
 
         full_end_time = timeit.default_timer()
-        self.logger.info(f"Time to Complete PR Eval: {full_end_time - full_start_time}")
+        self.logger.info(f"Time to Complete PR Eval: {(full_end_time - full_start_time):.3f} s")
 
         return 0