Adds `debug` level messages with the pattern "Time to" to log the time taken to perform important tasks. This provides greater insight into overall stuart execution time and allows results to be compared for regressions and across environments.

---

A build log can be filtered to "Time to" messages to get a condensed view of build time (from mu_tiano_platforms build):

```
Time to Get Git Version: 0.125 s
Time to Check If Repo is Dirty: 1.808 s
Time to Check for Worktrees: 2.413 s
Time to Gather Env Files: 5.288 s
Time to Bootstrap Environment: 11.939 s
Time to Verify Environment: 0.314 s
Time to Load Plugins: 1.205 s
Time to Start Invocable: 13.503 s
Time to Kick Off Platform Build: 7.243 s
Time to ParseTargetFile: 0.007 s
Time to ParseToolsDefFile: 0.036 s
Time to ParseDscFile: 0.908 s
Time to ParseFdfFile: 0.048 s
Time to SetEnv: 2.316 s
Time to run PlatformPreBuild: 0.014 s
Time to run do_pre_build() for Update Config Headers: 1145.764 ms
Time to run do_pre_build() for PECOFF Image Validation Plugin: 0.073 ms
Time to run do_pre_build() for Build Tools Report Generator: 4.783 ms
Time to run do_pre_build() for Clang Pdb Tool Chain Support: 0.162 ms
Time to run do_pre_build() for Debug Macro Check Plugin: 752.201 ms
Time to run do_pre_build() for Fd Size Report Generator: 0.003 ms
Time to run do_pre_build() for Flatten PDBs UEFI Build Plugin: 0.002 ms
Time to run do_pre_build() for Override Validation and Deprecation Warning Pre Build Plugin: 6608.771 ms
Time to run do_pre_build() for Windows RC Path Support: 1.816 ms
Time to run do_pre_build() for Windows Visual Studio Tool Chain Support: 3698.368 ms
Time to PreBuild: 12.239 s
Time to Build: 134.139 s
Time to run PlatformPostBuild: 0.000 s
Time to run do_post_build() for Update Config Headers: 0.057 ms
Time to run do_post_build() for PECOFF Image Validation Plugin: 5353.019 ms
Time to run do_post_build() for Build Tools Report Generator: 3.844 ms
Time to run do_post_build() for Clang Pdb Tool Chain Support: 0.084 ms
Time to run do_post_build() for Debug Macro Check Plugin: 0.002 ms
Time to run do_post_build() for Fd Size Report Generator: 247.504 ms
Time to run do_post_build() for Flatten PDBs UEFI Build Plugin: 875.959 ms
Time to run do_post_build() for Override Validation and Deprecation Warning Pre Build Plugin: 0.003 ms
Time to run do_post_build() for Windows RC Path Support: 0.002 ms
Time to run do_post_build() for Windows Visual Studio Tool Chain Support: 0.077 ms
Time to PostBuild: 6.494 s
```