@makubacki 2025-03-19
> Will push the separate logger change soon...

Joey will do this in a separate change - https://github.com/tianocore/edk2-pytool-extensions/issues/958.